PP-5072 - Move Apple Pay JS API stubs to utils
